+++ src/inet/applications/udpapp/UDPBasicApp.cc
@@ -115,7 +115,7 @@ void UDPBasicApp::sendPacket()
     std::ostringstream str;
     str << packetName << "-" << numSent;
     ApplicationPacket *payload = new ApplicationPacket(str.str().c_str());
-    payload->setByteLength(par("messageLength").longValue());
+    payload->setByteLength(par("messageLength").intValue());
     payload->setSequenceNumber(numSent);
 
     L3Address destAddr = chooseDestAddr();
