+++ src/inet/linklayer/lmac/LMacLayer.cc
@@ -141,7 +141,7 @@ InterfaceEntry *LMacLayer::createInterfaceEntry()
     e->setInterfaceToken(address.formInterfaceIdentifier());
 
     // capabilities
-    e->setMtu(par("mtu").longValue());
+    e->setMtu(par("mtu").intValue());
     e->setMulticast(false);
     e->setBroadcast(true);
