+++ src/inet/linklayer/ppp/PPP.cc
@@ -118,7 +118,7 @@ InterfaceEntry *PPP::createInterfaceEntry()
 
     // MTU: typical values are 576 (Internet de facto), 1500 (Ethernet-friendly),
     // 4000 (on some point-to-point links), 4470 (Cisco routers default, FDDI compatible)
-    e->setMtu(par("mtu").longValue());
+    e->setMtu(par("mtu").intValue());
 
     // capabilities
     e->setMulticast(true);
