+++ src/inet/transportlayer/tcp_common/TCPSpoof.cc
@@ -49,7 +49,7 @@ void TCPSpoof::sendSpoofPacket()
     int srcPort = par("srcPort");
     int destPort = par("destPort");
     bool isSYN = par("isSYN");
-    unsigned long seq = par("seqNo").longValue() == -1 ? chooseInitialSeqNum() : par("seqNo").longValue();
+    unsigned long seq = par("seqNo").intValue() == -1 ? chooseInitialSeqNum() : par("seqNo").intValue();
 
     // one can customize the following according to concrete needs
     tcpseg->setSrcPort(srcPort);
