+++ src/inet/transportlayer/tcp_nsc/TCP_NSC.cc
@@ -228,7 +228,7 @@ void TCP_NSC::initialize(int stage)
 
         // load the stack
         const char *stackName = this->par(stackNameParamNameS).stringValue();
-        int bufferSize = (int)(this->par(bufferSizeParamNameS).longValue());
+        int bufferSize = (int)(this->par(bufferSizeParamNameS).intValue());
         loadStack(stackName, bufferSize);
         pStackM->if_attach(localInnerIpS.str().c_str(), localInnerMaskS.str().c_str(), 1500);
         pStackM->add_default_gateway(localInnerGwS.str().c_str());
